latest calibration structure & code
