+++ lisa/calibrate.py
@@ -9,182 +9,19 @@ from collections import OrderedDict
 import sys
 from random import random, randrange, choice
 from nx_types import *
-
+import math
 from mire_check import *
 from ride_report_matching import *
+from calibrate_helper_functions import *
 
-
-
-def calc_distance_efficiency(edgenode1: NodeID, edgenode2: NodeID, endnode: NodeID):
-    """Takes in an edge and the endpoint to calculate how well the edge choice leads to the endpoint
-        Does this by calculating the distance between:
-            - first point to the end
-            - second point to the end
-            - length of the edge
-
-        returns ((first point to the end) - (second point to the end))/(length of edge)
-    Args:
-        edgenode1 (NodeID): Starting node identifier for the given edge
-        edgenode2 (NodeID): Ending node identifier for the given edge
-        endnode (NodeID): Final node identifier for the last node in the path
-
-    Returns:
-        efficiency_calculation (float): numerical representation of the progress towards the ending node
-    """
-    
-    (edgenode1_x, edgenode1_y) = G.nodes[edgenode1]['x'], G.nodes[edgenode1]['y']
-    (edgenode2_x, edgenode2_y) = G.nodes[edgenode2]['x'], G.nodes[edgenode2]['y']
-    (endnode_x, endnode_y) = G.nodes[endnode]['x'], G.nodes[endnode]['y']
-    
-    distance_edgenode1_edgenode2 = np.sqrt(np.square(edgenode1_x - edgenode2_x) + np.square(edgenode1_y - edgenode2_y))
-    distance_edgenode1_end = np.sqrt(np.square(edgenode1_x - endnode_x) + np.square(edgenode1_y - endnode_y))
-    distance_edgenode2_end = np.sqrt(np.square(endnode_x - edgenode2_x) + np.square(endnode_y - edgenode2_y))
-
-    progress_to_end = distance_edgenode1_end - distance_edgenode2_end
-    
-    efficiency_calculation = (progress_to_end/distance_edgenode1_edgenode2)
-    return efficiency_calculation
-
-def find_attribute_dict(G: nx.DiGraph, node1: NodeID, node2: NodeID, endnode: NodeID, featurelist: list):
-    """Returns the attributes of an edge from node1 to node2 in the context of G and endnode
-    
-    Args:
-        G (nx.DiGraph): DiGraph representation of the road network
-        node1 (NodeID): Starting node identifier for the given edge
-        node2 (NodeID): Ending node identifier for the given edge
-        endnode (NodeID): Node identifier of last node in the path for distance calculation
-        featurelist (list(tuples)): list of features that will be considered for model fitting
-
-    Returns:# the place to handle edge cases
-        attributes (dict): dictionary with an value for each feature in the featurelist
-    """
-    edgeID = (node1, node2)
-    current_edge_data = dict(G.edges[edgeID])
-    attributes = {}
-    
-
-    # TODO: see if we can restructure MIRE graph data so I don't write shitty code
-    for feature in featurelist:
-        # software design is hard...
-        if (feature == "distance_efficiency"):
-            feature_value = calc_distance_efficiency(node1, node2, endnode)
-
-        # elif (feature == "has_comp"):
-        #     assert (len(list(G.neighbors(node2))) == 1), "There are too many neighbors to get has_comp"
-        #     neighbor_of_neighbor = list(G.neighbors(node2))[0]
-        #     next_edge_id = (node2, neighbor_of_neighbor)
-        #     next_edge_data = dict(G.edges[next_edge_id])
-        #     feature_value = next_edge_data["has_comp"]
-
-        else:
-            try:
-                feature_value = current_edge_data['attributes'][feature]
-            except Exception as e:
-                print(e)
-                print("Feature " + feature + " not found")
-                print("attribute dictionary: ", current_edge_data["attributes"])
-                print("edge ID: ", edgeID)
-                feature_value = False
-        
-        attributes[feature] = feature_value
-    assert (len(attributes) == len(featurelist)), "attributes length: " + str(len(attributes)) + "featurelist length: " + str(len(featurelist))
-    return attributes
-
-
-    
-def create_dataframe(G: nx.DiGraph, paths: list, feature_list: list):
-    """Create the pandas dataframe that is compatible with pylogit
-    
-    Args:
-        G (nx.DiGraph): DiGraph representation of the road network
-        paths (list[NodeID]): chosen list of nodeIDs the biker visited
-        feature_list (list[str]): list of (id#, featurename) tuples present in G.edges output
-    
-    Returns:
-        (dataframe, specs, spec_names): Pandas dataframe and spec settings needed to run pylogit model
-    """
-    n_feats = len(feature_list)
-    choice_features = []
-    observation_ids = []
-    choice_indicators = []
-    choice_ids = []
-    observation_id = 0
-
-    for path_index in range(len(paths)):
-        
-        path = paths[path_index]
-
-        for i in range(len(path)-1):
-
-            # append all the edge options to a list of lists called choice_features
-            current_node = path[i]
-            end_node = path[-1]
-            neighbors = list(G.neighbors(current_node))
-            n_choices = len(neighbors)
-
-
-            # attempting to add to the intersections dataframe ONLY if # choices > 1
-            if (n_choices > 1):
-                
-                observation_ids.append(observation_id*np.ones((n_choices,))) 
-                # 'i' is the "index" of the observation, or the reason why we know which observation is which
-                for neighbor in neighbors:
-                    
-                    edgeID = (current_node, neighbor)
-
-                    type = dict(G.edges[edgeID])['type'] # line that might be useful for later
-                    
-                    current_attribute_dict = find_attribute_dict(G, current_node, neighbor, end_node, feature_list)
-                    
-                    # collect all the observations for all neighbors
-                    current_observation_choice_features = []
-
-                    # iteratively adding each feature value to the observation
-                    for feature in feature_list:
-                        current_feature = current_attribute_dict[feature]
-                        current_observation_choice_features.append(current_feature)
-                    
-                    # append each possibility's features to the dataframe
-                    choice_features.append(current_observation_choice_features)
-
-                # marking the choiceID's choice as '1' among zeros
-                choice_indicators.append(np.zeros((n_choices,)))
-                chosen = neighbors.index(path[i+1])
-                choice_indicators[-1][chosen] = 1
-                # All the possible choices out at this observation:
-                choice_ids.append(np.arange(n_choices))
-                
-                observation_id += 1
-    
-    # preparing columns for the dataframe (long) format
-    overall_observation_ids = np.concatenate(observation_ids)
-    choice_features_overall = np.vstack(choice_features)
-    overall_choice_indicators = np.concatenate(choice_indicators)
-    overall_choice_ids = np.concatenate(choice_ids)
-
-    df_intersections = pd.DataFrame()
-    
-    df_intersections['obs_ids'] = overall_observation_ids
-    df_intersections['choices'] = overall_choice_indicators
-    df_intersections['alt_ids'] = overall_choice_ids
-    
-    
-    # The next few lines just mean the columns will be consistent across choices
-    spec_names = OrderedDict()
-    specs = OrderedDict()
-    for i in range(n_feats):
-        spec = feature_list[i]
-        spec_names[spec] = spec
-        specs[spec] = 'all_same'
-        df_intersections[spec] = choice_features_overall[:,i]
-
-    return (df_intersections, specs, spec_names)
-
-def create_model(dataframe: pd.DataFrame, specs: OrderedDict, spec_names: OrderedDict):
+def create_model(dataframe: pd.DataFrame, feature_list: list):
     """Fitting multinomial logit model to the dataframe choices with specifications
     
     Args:
         dataframe (pd.DataFrame): Dataframe representation of the routes and the choices taken in our data
+        feature_list (list[str]): list of (id#, featurename) tuples present in G.edges output
+        
+        
         specs (OrderedDict): Each category with an 'all_same' value to make sure the column ID does not change for the same attribute
         spec_names (OrderedDict): Customizable specs for each category for nesting cases
     
@@ -192,9 +29,16 @@ def create_model(dataframe: pd.DataFrame, specs: OrderedDict, spec_names: Ordere
         (fit_summary, summary): fit_summary is a dictionary that shows the overall model's fit, 
                                 summary is a dictionary that shows each input's fit in the model
     """
-    assert (len(specs) == len(spec_names))
     
-    n_feats = len(specs)
+    # This just means the columns will be consistent across choices
+    spec_names = OrderedDict()
+    specs = OrderedDict()
+    for i in range(len(feature_list)):
+        spec = feature_list[i]
+        spec_names[spec] = spec
+        specs[spec] = 'all_same'
+
+    assert (len(specs) == len(spec_names))
 
     # Fit to a multinomial logit model (MNL)
     choice_model = pl.create_choice_model(data=dataframe,
@@ -205,8 +49,7 @@ def create_model(dataframe: pd.DataFrame, specs: OrderedDict, spec_names: Ordere
                                             model_type="MNL",
                                             names=spec_names)
     
-    
-    choice_model.fit_mle(np.zeros(n_feats), print_res = False)
+    choice_model.fit_mle(np.zeros(len(feature_list)), print_res = False)
     
     fit_summary_print_output = choice_model.fit_summary
     summary_print_output = choice_model.summary
@@ -216,9 +59,6 @@ def create_model(dataframe: pd.DataFrame, specs: OrderedDict, spec_names: Ordere
     fit_summary = (fit_summary_print_output.to_dict())
 
     choice_model.get_statsmodels_summary()
-
-    print(fit_summary_print_output)
-    print(summary_print_output)
 
     return fit_summary, summary
 
@@ -235,7 +75,7 @@ def sort_input_attributes(summary_dict: dict):
     """
     weights = {}
 
-    # try to reject any values p_values greater than .05
+    # try to reject any variables not worth considering
     for attribute in summary_dict:
         std_error = summary_dict[attribute]['std_err']
         weight = summary_dict[attribute]['parameters']
@@ -257,14 +97,8 @@ def weights_to_lts(attribute_weights: dict):
     
 
 
-
 if __name__ == "__main__":
 
-    # bbox = Bbox(38.88300016, 38.878726840000006, -77.09939832, -77.10500768)
-    # G = Graph(bbox)
-    # G.save("boundgraph")
-
-    # # G = Graph.from_file("dc.pickle").DiGraph
     gdb = "scratch_022819.gdb"
     node_layer = 3
     edge_layer = 2
@@ -285,17 +119,17 @@ if __name__ == "__main__":
 
         except nx.exception.NetworkXNoPath as e:
             print(e)
-        
-    print("res: ", res)
     
     intersections_featurelist = ['distance_efficiency', 'notAtGrade', 'stops', 'signal', 'pedsignal', 'rr', 'yield']
     featurelist = ['distance_efficiency', 'has_comp', 'notAtGrade', 'stops', 'signal', 'pedsignal', 'rr', 'yield']
 
-    (df, specs, spec_names) = create_dataframe(G, res, intersections_featurelist)
-    print("dataframe\n", df[:15])
-    fit_summary, summary_dict = create_model(df, specs, spec_names)
-    # # print(a)
-    # # print(b)
-    # print(sort_input_attributes(summary_dict))
+    (df_intersections, _, _) = create_dataframes(G, res, intersections_featurelist)
+    print("dataframe:\n", df_intersections[:15])
+    fit_summary, summary_dict = create_model(df_intersections, intersections_featurelist)
+    
+    import pprint
+    pprint.pprint(fit_summary)
+    pprint.pprint(summary_dict)
+    print(sort_input_attributes(summary_dict))
